Merge main into 5.x.y (#29)
* :construction_worker: Migrate to clang for host tests

* :arrow_up: Set deploy to C++23

* :sparkles: Add deploy_linux

deploy_linux.yml is used for header only libraries and CI to be able to
match the github actions OS and settings to what is used in
conan-config.

Update README.md to be mostly up to date.

* :sparkles: Add support for external packages

deploy.yml will now accept an "external_package" input which, when set
to 'true' will clone the repo from main and not from a release tag
reference. This is useful for projects where we are just supporting
a conan package of a package and not creating the releases ourselves.