:construction_worker: Remove conan create for demo builds
I realize that it was a mistake to build the current packages before
building demos. Devs will attempt to clone our repo and build our demos,
but may fail if a new demo using new features is added to the list. To
ensure that this cannot happen, the current package is not built and the
demo must be built using whats available on the conan repo.